backend/pyproject.toml
[project]
name = "edos-security-backend"
version = "1.0.0"
description = "EDoS Security Dashboard Backend - FastAPI backend for real-time cybersecurity monitoring"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "bcrypt>=5.0.0",
    "email-validator>=2.3.0",
    "faker>=38.2.0",
    "fastapi>=0.121.3",
    "keras>=3.12.0",
    "lightgbm>=4.6.0",
    "loguru>=0.7.3",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.0",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "scikit-learn>=1.7.2",
    "sqlalchemy>=2.0.44",
    "supabase>=2.24.0",
    "tensorflow[and-cuda]>=2.20.0",
    "uvicorn>=0.38.0",
    "xgboost>=3.1.2",
    "redis>=5.0.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app", "ml"]

